ark_collector.cpp告警处理

+++ common_components/heap/ark_collector/ark_collector.cpp
@@ -708,7 +708,7 @@ void ArkCollector::ParallelFixHeap()
     auto taskList = regionSpace.CollectFixTasks();
     std::atomic<int> taskIter = 0;
     std::function<FixHeapTask *()> getNextTask = [&taskIter, &taskList]() -> FixHeapTask* {
-        uint32_t idx = taskIter.fetch_add(1U, std::memory_order_relaxed);
+        uint32_t idx = static_cast<uint32_t>(taskIter.fetch_add(1U, std::memory_order_relaxed));
         if (idx < taskList.size()) {
             return &taskList[idx];
         }
